backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
PyMuPDF==1.23.8
reportlab==4.0.7
pypdf==3.17.1
Pillow==10.1.0
python-jose[cryptography]==3.3.0
aiofiles==23.2.1
